Rename LCD lib
